ENH Refactor link field traits into abstract class
